# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Tienda",
        "description": "Esta es la documentación de la API que permite gestionar categorías, productos y órdenes en una tienda.",
        "contact": {
            "email": "jmcabo93@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtiene una lista de productos",
                "description": "Devuelve una lista de productos paginados.",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Lista de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron productos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crea un nuevo producto",
                "description": "Almacena un nuevo producto en la base de datos.",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtiene un producto específico",
                "description": "Devuelve los detalles de un producto por su ID.",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualiza un producto existente",
                "description": "Actualiza los datos de un producto por su ID.",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Elimina un producto",
                "description": "Elimina un producto de la base de datos por su ID.",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Producto eliminado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/random": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtiene un producto aleatorio",
                "description": "Devuelve un producto aleatorio de la base de datos.",
                "operationId": "4f3d8bf6285548db3805678002f3220d",
                "responses": {
                    "200": {
                        "description": "Producto aleatorio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay productos disponibles"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Product": {
                "required": [
                    "name",
                    "description",
                    "price",
                    "stock",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID del producto",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del producto",
                        "type": "string",
                        "example": "Smartphone"
                    },
                    "description": {
                        "description": "Descripción del producto",
                        "type": "string",
                        "example": "Electronics"
                    },
                    "price": {
                        "description": "Precio del producto",
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "stock": {
                        "description": "Cantidad de stock disponible",
                        "type": "integer",
                        "example": 200
                    },
                    "category_id": {
                        "description": "ID de la categoría a la que pertenece el producto",
                        "type": "integer",
                        "example": 1
                    },
                    "image": {
                        "description": "Imagen del producto",
                        "type": "string",
                        "example": "",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usa un token JWT en formato Bearer para autenticarte",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}